requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
matplotlib==3.9.2
PyYAML==6.0.2
torch==2.2.2
tqdm==4.66.5
pytest==8.3.3
transformers
huggingface_hub

# Optional for the original TimeOmni benchmark pipeline in timeomni_core/eval/
# vllm
